Restructure logic instance/proof: move LogicVerifierInputs, add to_journal
- Move LogicVerifierInputs from logic_proof.rs to logic_instance.rs
- Add to_journal() on LogicInstance gated on zkvm
- Add get_inner_receipt() on LogicVerifierInputs gated on zkvm
- Use instance.to_journal() in to_logic_verifier() instead of risc0_zkvm::serde::to_vec
- Import LogicVerifierInputs from logic_instance in logic_proof.rs
