+++ arm/src/logic_proof.rs
@@ -3,16 +3,15 @@
 use crate::{
     constants::{PADDING_LOGIC_PK, PADDING_LOGIC_VK},
     error::ArmError,
-    logic_instance::{AppData, LogicInstance},
+    logic_instance::{LogicInstance, LogicVerifierInputs},
     nullifier_key::{NullifierKey, NullifierKeyCommitment},
     proving_system::{journal_to_instance, verify as verify_proof},
     resource::Resource,
     resource_logic::TrivialLogicWitness,
-    utils::words_to_bytes,
 };
 use rand::rngs::OsRng;
 use rand::Rng;
-use risc0_zkvm::{serde::to_vec, sha::Digest, InnerReceipt};
+use risc0_zkvm::sha::Digest;
 use serde::{Deserialize, Serialize};
 
 #[cfg(feature = "prove")]
@@ -60,19 +59,6 @@ pub struct LogicVerifier {
     pub verifying_key: Digest,
 }
 
-/// Inputs required to create a logic verifier.
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq)]
-pub struct LogicVerifierInputs {
-    /// The tag (either commitment or nullifier) for the logic instance.
-    pub tag: Digest,
-    /// The verifying key for the logic proof.
-    pub verifying_key: Digest,
-    /// The application data associated with the logic instance.
-    pub app_data: AppData,
-    /// The logic proof (optional, would be absent when aggregation is enabled).
-    pub proof: Option<Vec<u8>>,
-}
-
 impl LogicVerifier {
     /// Verifies the logic proof against the instance using the provided verifying key.
     pub fn verify(&self) -> Result<(), ArmError> {
@@ -99,34 +85,12 @@ impl LogicVerifierInputs {
         is_consumed: bool,
         root: Digest,
     ) -> Result<LogicVerifier, ArmError> {
-        let instance_words = to_vec(&self.to_instance(is_consumed, root))
-            .map_err(|_| ArmError::InstanceSerializationFailed)?;
+        let instance = self.to_instance(is_consumed, root);
         Ok(LogicVerifier {
             proof: self.proof,
-            instance: words_to_bytes(&instance_words).to_vec(),
+            instance: instance.to_journal()?,
             verifying_key: self.verifying_key,
         })
-    }
-
-    /// Converts the LogicVerifierInputs into a LogicInstance.
-    fn to_instance(&self, is_consumed: bool, root: Digest) -> LogicInstance {
-        LogicInstance {
-            tag: self.tag,
-            is_consumed,
-            root,
-            app_data: self.app_data.clone(),
-        }
-    }
-
-    /// Retrieves the inner receipt from the logic proof.
-    pub fn get_inner_receipt(&self) -> Result<InnerReceipt, ArmError> {
-        let inner: InnerReceipt = bincode::deserialize(
-            self.proof
-                .as_ref()
-                .ok_or(ArmError::MissingField("Missing logic proof"))?,
-        )
-        .map_err(|_| ArmError::InnerReceiptDeserializationError)?;
-        Ok(inner)
     }
 }
